Augmented the code with unlabeled data.
After each epoch on 1464 labeled training data, there will be an epoch over 8498 unlabed data.
For each image in the unlabled set, a pseudo-label is estimated and uased as its target for optimization.
Right now they are not used be vause of an If(0) condition in train_epoch() function.
So the output of this commit is similar to the output of the previous commit.