train_siamfcn32s tries to train a binary Siamese-FCN network.
The network is fed with a pair of images (of the same sizes. in this implementation it is 256x256, for which, the transform function in voc.py has been modified) and predicts for each pixel whether or not two images have the same semantic label.

NOTE: By modifying the transform function in voc.py, the outputs of train_fcn32.py is changed, as we have resized images and label maps.

So far, I have only written the basic code that reads the training data and for each training image, searches through the rest of the training data and for the ones with sufficient simlarity, a label compatibility loss is computed. In other words, the ground truth for the final output of the siamese-fcn network is one for the pixels that have identical labels in both image, and zero for the corresponding pixels that belong to different classes.

So far, by analyzing for a few images, I have seen that cosine_similarity metric on image features does a good job on finding similar images. Of course the similarity threshold sould be optimized through validation. Right now it is opted to be 0.5.

Once a similar input pair are identified, they pass through the the Merge class which is applied to the features of model-FCN32 (self.model). In this layer, I have used a simple inner-product and then a simple binary classifier(segmenter), but I am not sure if the inner-product is a better choice than Euclidan distance. Have not gone through analysis of this stage.