Training a binary FCN-Siamese network.
For each training sample, all other training samples are checked and the similar ones are used for training the model.

Note that for similarity checking, the fixed pre-trained FCN32s model is used and it is is not updated during the training of Siamese network.

Also note that you are redundantly computing the similarity feature vectors over and over again which makes the training much slower.
You'd better compute the feature vectors for all 1464 images once and then use them throughout the training.

Also note I do not expect a very good result from this Pascal dataset and I am going to implement dataloder for SiftFlow.