Clustering loss was added.
First the mean_embedding for each class is computed using get_mean_embeddings.py and a pre-trained network on that had an accuracy of 58%. The embeddings are obtained from the 21D layer before softmax.
Then these mean_embeddings are considerred as cluster centers for clustering loss.
For validation, a nearest neighbour search between the embedding of each pixel and the mean_embeddings is performed. Validation starts from 41% in the first validation check, and goes up to 48% and even though it is very slow, it almost keeps improving.
The negative point is that the mean_embeddings that are obtained from a network with 58% performance, after the first 4000 iteration of the clustering loss, led to 41% and even though it gets better, these is a big gap to 58%.

Note that I have also changed the cross-entropy loss a little bit, which I am going to revert it back to its previous version later.