Just refined the trainer.py.
Note that the train_dataloder in train_fcn32s.py is defined with SHUFFLE=False. We did this to make sure if we set aside some images from the 1449 training data to clustering loss, then in the next epoch the very same images are passed to the clustering loss.

Also defined new optimizer for clustering loss.
However the clustering loss did not work well and it probably needs to be changed.
For now given the condition in traner.py: line 218, we never use the clustering loss and training should work like the previous versions (e.g. 376c532) with a mean-IOU=45~48% after iteration=4000.