First commit for DCSP in Pytorch.
After training for one epoch, it seems to work fine and does a fair localization for the first few validation images.

In this first version, I have defined a wTrainer for weak training.
Note that I have considered two cases (Pascal or non-Pascal).
In the case of Pascal dataset (in the code, whenever n_class==21), the loss is computed on the (n_class-1) score maps (score_w), which excludes background.
However for other cases (e.g. Camvid, etc.) loss is computed normally for the n_class dimensional layer.
Because of this discrepancy, I have modified fcn32s.py to get two n_class and (n_class-1) dimensional outputs.
Another importanct change in fcn32s compared to the previous commit is in the _initialize_weights function. Zero initialization of the weights corrupted the training process, so I changed it to Xavier.

Also note that the dataloader now outputs three items: {Image, Target, Tags}.
For VOC data, the Tags output excludes Background.

Note that in train_fcn32s, wTrainer is called instead of Trainer.
Also please note the learning rate is changed significantly, I did not play with it after changing it to 1e-6.

Also note that inside wTrainer at the constructer, we freezed the parameters of the Convlutional layers before fc layers, as done in Arslan's implementation.
Have not tested their impact.

Sigmoid with Binary cross entropy is chosen for computing the loss function.

Note that this is only the Attention part of the paper and I have not used any Saliency cues. Besides, I have discarded the min-max normalization as done in Arslan's implementation as it gave me poor results for the first validation images. When I changed this min-max-normalization to normalization in the direction of the classes, I got much better results.