Making the weakly and fully-supervised codes consistent. Just remember to use Caffe_to_numpy.py code to have a proper pre-trained model. (Refer to previous commits for more information on caffe_to_numpy)
The attention-based weakly supervised model leads to 24% m-IoU after one epoch and does not get better after that.

This prediction should be used as pseudo-GT for later iterations. Though some confidence measures should be taken into account for this localization cue, as it is not reliable as a dense GT.