Added class Attention model.
This was just a very simple solution for having different weight initializations for the attention layer and FCN layers.
Because of this, now two models are processed and saved in the framework: (model, model_att).
However I do not like this structure and I will merge them into one single model in the next commits.

Then in order to solve the problem of different scales of weight initialization, I do a better thing by assigning a learning rate to attention layer, which is 1000 scales larger than the learning rates of fc layers. (Note that in the weak framework, the feature layers are frozen)
So it would be for instance this way for weak supervision:
Normal learning rates for fc layers: 1e-5
For attention layer: 1e-2