Localization and weak segmentation was added.
get_localization.py estimates label localizations given the "trained attention model using wTrainer".
For full supervision using these generated pseudo ground truth cues, torchfcn.tagSupTrainer is used, which is almost the same as torchfcn.Trainer, with this difference that for each training sample, its original ground truth is replaced by the estimated ground truth.

Note that it does not work well at all and this commit is just for saving what has been done so far. There are many hyper parameters like thresholds for localization and so on that need to be adjusted. It is a bit ugly!